h5pt: test explicit intXY

+++ h5pt_test.go
@@ -13,13 +13,13 @@ const (
 )
 
 type particle struct {
-	// name        string  `hdf5:"Name"`     // FIXME(sbinet)
-	lati        int     `hdf5:"Latitude"`
-	longi       int     `hdf5:"Longitude"`
+	// name        string  `hdf5:"Name"`      // FIXME(sbinet)
+	lati        int32   `hdf5:"Latitude"`
+	longi       int64   `hdf5:"Longitude"`
 	pressure    float32 `hdf5:"Pressure"`
 	temperature float64 `hdf5:"Temperature"`
 	// isthep      []int                     // FIXME(sbinet)
-	// jmohep      [2][2]int                 // FIXME(sbinet)
+	// jmohep [2][2]int64                    // FIXME(sbinet)
 }
 
 func TestTable(t *testing.T) {
